+++ packages/cli/templates/generate/intent/{{intentName}}.ts
@@ -1,5 +1,5 @@
 import { {{intentType}}, TContext } from '@bearer/intents'
-import Client from './Client'
+import Client from './client'
 
 export default class {{intentName}}Intent {
   static intentName: string = '{{intentName}}'