+++ packages/transpiler/bin/bearer-tst.ts
@@ -2,10 +2,9 @@
 import Transpiler from '../src/index'
 
 export default args => {
-  const transpiler = new Transpiler(
-    undefined,
-    args.indexOf('--no-watcher') === -1
-  )
+  const transpiler = new Transpiler({
+    watchFiles: args.indexOf('--no-watcher') === -1
+  })
 
   transpiler.on('STOP', () => {
     process.exit(0)