+++ packages/templates/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.20.0"></a>
+# [0.20.0](https://github.com/Bearer/bearer/compare/v0.19.1...v0.20.0) (2018-07-20)
+
+
+### Features
+
+* **templates:** Add BasicAuth and update Client ([#120](https://github.com/Bearer/bearer/issues/120)) ([afc0192](https://github.com/Bearer/bearer/commit/afc0192))
+
+
+
+
 <a name="0.18.0"></a>
 # [0.18.0](https://github.com/Bearer/bearer/compare/v0.17.5...v0.18.0) (2018-07-19)