+++ packages/ui/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.20.0"></a>
+# [0.20.0](https://github.com/ionic-team/stencil-component-starter/compare/v0.19.1...v0.20.0) (2018-07-20)
+
+
+
+
+**Note:** Version bump only for package @bearer/ui
+
 <a name="0.19.0"></a>
 # [0.19.0](https://github.com/ionic-team/stencil-component-starter/compare/v0.18.0...v0.19.0) (2018-07-19)